requirements.txt
pandas
pywin32
python-dotenv